Avoiding errors in url
Added a pre-condition in `preferred_forms_check` before appending errors to check whether the first character is any alphabet or not, because the `regex` search returns the match with non-word characters also on both sides of the `input match`.
Say, for example the input search string is `june` and it has to be searched in the following text " This page shows results of month june, https://www.google.com/search?q=june ". Then word june in url will be returned with `=` also, and this character is checked whether the returned match contains any non word character or not.